move utility functions to module
